# openAPI.json
openapi: "3.0.0"
info:
  title: WhatsApp API MultiDevice
  version: 8.3.1
  description: |
    This API is used for sending whatsapp via API.

    Device scoping:
    - Send `X-Device-Id` on all device-scoped REST calls.
    - WebSocket: connect to `/ws?device_id=<id>`.
servers:
  - url: https://gowa-vl0g.srv909561.hstgr.cloud
...
security:
  - basicAuth: []
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags:
        - device
      summary: List all devices
      description: Returns all registered devices with their connection status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    post:
      operationId: addDevice
      tags:
        - device
      summary: Add a new device
      description: Create a new device slot for multi-device management
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
                  description: Optional custom device ID. If not provided, one will be generated.
                  example: 'my-custom-device-id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAddResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    DeviceListResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: List devices
        status:
          type: integer
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeviceInfo'
    DeviceAddResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Device added
        status:
          type: integer
          example: 200
        results:
          $ref: '#/components/schemas/DeviceInfo'
...
    DeviceInfo:
      type: object
      properties:
        id:
          type: string
          example: 'my-device-id'
        phone_number:
          type: string
          example: '628123456789'
        display_name:
          type: string
          example: 'John Doe'
        state:
          type: string
          enum: [disconnected, connected, logged_in]
          example: 'logged_in'
        jid:
          type: string
          example: '628123456789@s.whatsapp.net'
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
...
    ErrorInternalServer:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
          description: 'SYSTEM_CODE_ERROR'
        message:
          type: string
          example: you are not loggin
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
    ErrorBadRequest:
      type: object
      properties:
        code:
          type: string
          example: 400
          description: 'HTTP Status Code'
        message:
          type: string
          example: field cannot be blank
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
...